==> portfolio/analysis/4a-createData_linear_modeling.Rmd <==
---
title: "Create data sets for modeling"
output:
  pdf_document:
    fig_caption: yes
  html_document: default
---

## Differenced vs. undifferenced datasets:

1. Excluding observations that we don't have consecutive years for

```{r setup, include=FALSE, echo=FALSE}
knitr::opts_chunk$set(echo = TRUE, tidy=FALSE, tidy.opts=list(width.cutoff=60), warning = FALSE, message = FALSE)
```

## Load cleaned data

```{r dataLoad, warning = FALSE, message=FALSE, results="hide", echo=FALSE}
library(dplyr)
library(knitr)
library(ggplot2)
library(date)

if (Sys.info()[["user"]] == "seananderson") {
  cfec <- feather::read_feather("../data-generated/cfec.feather")
} else {
  cfec <- feather::read_feather("data/cfec.feather")
}

# convert 6-dig p_fshy codes to 5 digits
# removing blank space - now they're all length(5)
idx = which(nchar(cfec$p_fshy)==6)
cfec$p_fshy[idx] = paste(substr(cfec$p_fshy[idx], 1, 1),
  substr(cfec$p_fshy[idx], 3, 6), sep="")

# remove white space in 2nd position
idx = which(substr(cfec$p_fshy, 2, 2) == " ")
cfec$p_fshy[idx] = paste(substr(cfec$p_fshy[idx], 1, 1),
  substr(cfec$p_fshy[idx], 3, length(cfec$p_fshy[idx])), sep="")

if (Sys.info()[["user"]] != "seananderson") {
  region<-read.csv("areas/regions.csv")
  region<-region[,c(1:2)] #don’t include “assigned by” column
  cfec<-left_join(cfec,region)
}
```

We need to adjust the landed values and prices for inflation. To do this,
we'll use the data Alan pulled together (adjusting to 2009 dollars).

<https://docs.google.com/spreadsheets/d/1hRLYTX1nZBLKnYXwTCOtU7-XnDoMZf_1ufdTiNQFd4U/edit#gid=0>

```{r, echo=FALSE}
if (Sys.info()[["user"]] != "seananderson") {
deflationTable = data.frame(year = 1985:2014, 
  defl = c(0.57236, 0.58393, 0.59879, 0.61974,
0.64388, 0.66774, 0.68993, 0.70564, 0.72244, 0.73781,
0.75321, 0.76695, 0.78009, 0.78855, 0.80061,
0.81883, 0.83753, 0.85038, 0.86729, 0.89114,
0.91981, 0.94812, 0.97334, 0.9925, 1,
1.01217, 1.03307, 1.05213, 1.06926, 1.08682))

# Adjust price and g_earn for inflation
cfec$year = as.numeric(cfec$year)
cfec$day = substr(cfec$landdate, 6, 10)
cfec$g_price = cfec$g_price/deflationTable$defl[cfec$year-1984]
cfec$g_earn = cfec$g_earn/deflationTable$defl[cfec$year-1984]
}
```

## Function definition

We need to define a few functions for calculating diversity. We'll use the
Simpson's diversity index (HHI) to calculate the entropy-based inverse, which
is a proxy for effective species diversity.

```{r, warning = FALSE, message=FALSE, echo=FALSE}
simp.div = function(x) {
  1/sum((x/sum(x))^2)
}
```

## Filtering out unique vessel - person combinations. Drops rows from 17 to 13 million.

Now we'll filter out (i.e. keep) individuals who only used 1 boat in a year, and boats with
only one permit holder fishing in a year.

```{r}
cfecnew = cfec
cfec.1boat = group_by(cfecnew, p_holder, year) %>%
  summarize(nBoat = length(unique(cadfg))) %>%
  filter(nBoat==1) %>%
  left_join(cfecnew)

cfecnew = cfec.1boat
dim(cfecnew)[1]
```

## Combine some strategies 

```{r}
cfecnew$p_fshy[cfecnew$p_fshy=="B61B"] = "B06B"
cfecnew$p_fshy[cfecnew$p_fshy=="C61B"] = "C06B"
cfecnew$p_fshy[cfecnew$p_fshy=="C91B"] = "C09B"
cfecnew$p_fshy[cfecnew$p_fshy=="M61B"] = "M06B"
cfecnew$p_fshy[cfecnew$p_fshy=="E09B"] = "E91B"
cfecnew$p_fshy[cfecnew$p_fshy=="I06B"] = "I61B"
cfecnew$p_fshy[cfecnew$p_fshy=="Y06A"] = "Y61A"
cfecnew$p_fshy[cfecnew$p_fshy=="O06B"] = "O61B"
cfecnew$p_fshy[cfecnew$p_fshy=="O09B"] = "O91B"

cfecnew$p_fshy[cfecnew$p_fshy%in%c("M7FB","M7GB", "M7HB","M7IB")] = "M7IB"
cfecnew$p_fshy[cfecnew$p_fshy%in%c("M7FG","M7GG", "M7HG","M7IG")] = "M7IG"

# crab and shrimp vessels
cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="D91")] = paste("D09", substr(cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="D91")],4,length(cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="D91")])), sep="")
cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="T91")] = paste("T09", substr(cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="T91")],4,length(cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="T91")])), sep="")
cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="P91")] = paste("P09", substr(cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="P91")],4,length(cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="P91")])), sep="")

cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="M91")] = paste("M09", substr(cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="M91")],4,length(cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="M91")])), sep="")
cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="M61")] = paste("M06", substr(cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="M61")],4,length(cfecnew$p_fshy[which(substr(cfecnew$p_fshy,1,3)=="M61")])), sep="")
```

## Identify most popular strategies by person-year combination

Any person in a year will need to have 10% or more of their revenue to come
from each permit (otherwise doesn't count as a strategy).

The top 120 permits cuts the rows down to 8.6 million. The top 50 gets it down
to 7.8 million.

```{r}
permit_rev = group_by(cfecnew, p_holder, year, p_fshy) %>%
  summarize(permit.rev = sum(g_earn)) %>%
  as.data.frame() %>%
  group_by(p_holder, year) %>%
  mutate(tot.rev = sum(permit.rev))
permit_rev$permit.rev = permit_rev$permit.rev/permit_rev$tot.rev # fraction of total
permit_rev = permit_rev[permit_rev$permit.rev < 0.1,]

# Delete these minimal earnings
cfecnew = cfecnew[-which(paste(cfecnew$p_holder, cfecnew$year, cfecnew$p_fshy) %in% paste(permit_rev$p_holder, permit_rev$year, permit_rev$p_fshy)), ]

topstrategies = group_by(cfecnew, p_holder, year) %>%
  summarize(strategy = paste(sort(unique(p_fshy)), collapse = " ")) %>%
  as.data.frame()
cfecnew = left_join(cfecnew, topstrategies) # adds strategy column

write.csv(rev(sort(table(topstrategies$strategy))), file = "../data-generated/strategies.csv")
```

Aggregate annual metrics by person,

```{r}

cfecnew$yday <- paste(cfecnew$year, cfecnew$day, sep = "-") %>%
  lubridate::ymd() %>%
  lubridate::yday() - 1
  
# fixme What is cadfg2?
if ("VLength" %in% names(cfecnew)) {
  cfecnew <- rename(cfecnew, vlength = VLength)
}
cfecAnnual = group_by(cfecnew, p_holder, year) %>%
  summarize(permit = p_fshy[1], boat = cadfg[1], # fixme this was cadfg2
    days = diff(range(yday) + 1),
    ngear = length(unique(gearn)),
    nspec = length(unique(specn)),
    npermit = length(unique(p_fshy)),
    revenue = sum(g_earn),
    weight = sum(g_pounds),
    length = mean(vlength, na.rm=TRUE),
    strategy = strategy[1])

# also calculate weighted simpson diversity of gears and species
cfecAnnual = left_join(cfecAnnual, group_by(cfecnew, p_holder, year, specn) %>%
  summarize(totRev.spec = sum(g_earn)) %>%
  group_by(p_holder, year) %>%
  summarize(specDiv = simp.div(totRev.spec)))

# also calculate weighted simpson diversity of gears and species
cfecAnnual = left_join(cfecAnnual, group_by(cfecnew, p_holder, year, gearn) %>%
  summarize(totRev.gear = sum(g_earn)) %>%
  group_by(p_holder, year) %>%
  summarize(gearDiv = simp.div(totRev.gear)))

# Also merge in previous revenue. Idx stores index of people last year
idx = match(paste(cfecAnnual$p_holder,cfecAnnual$year-1), paste(cfecAnnual$p_holder,cfecAnnual$year))

cfecAnnual$permit.prev = cfecAnnual$permit[idx]
cfecAnnual$ngear.prev = cfecAnnual$ngear[idx]
cfecAnnual$nspec.prev = cfecAnnual$nspec[idx]
cfecAnnual$npermit.prev = cfecAnnual$npermit[idx]
cfecAnnual$revenue.prev = cfecAnnual$revenue[idx]
cfecAnnual$specdiv.prev = cfecAnnual$specDiv[idx]
cfecAnnual$geardiv.prev = cfecAnnual$gearDiv[idx]
cfecAnnual$strategy.prev = cfecAnnual$strategy[idx]
cfecAnnual$weight.prev = cfecAnnual$weight[idx]
cfecAnnual$days.prev = cfecAnnual$days[idx]

# idx = which(cfecAnnual$strategy != cfecAnnual$strategy.prev)
```

Save the data for modeling:

```{r}
# make sure the data are tidy:
cfecAnnual <- as_data_frame(cfecAnnual)

## Filtering out people with 10+ years of fishing. Drops rows from 13.7 to 10 million.
## (now doing this at the beginning of the modeling scripts)
# cfecAnnual = group_by(cfecAnnual, p_holder) %>%
#   mutate(nyr = length(unique(year))) %>%
#   filter(nyr >= 10)

saveRDS(cfecAnnual, file="../data-generated/cfec-annual-for-modeling.rds")

cfecAnnual$specDiv.change = cfecAnnual$specDiv - cfecAnnual$specdiv.prev
cfecAnnual.diff = cfecAnnual[is.na(cfecAnnual$specDiv.change)==FALSE, ]
saveRDS(cfecAnnual.diff, file="../data-generated/cfec-diff-for-modeling.rds")
```

## Do some exploration by permit types + diversity

For example, when someone is diversifying with a herring, halibut, or crab permit, what does that diversification look like?

```{r, echo=FALSE}
cfecnew$taxa = substr(cfecnew$p_fshy, 1, 1)
specPermit = group_by(cfecnew, taxa, specn) %>%
  summarize(percent.revenue = sum(g_earn))
specPermit.tot = group_by(specPermit, taxa) %>%
  summarize(tot = sum(percent.revenue))
specPermit = left_join(specPermit, specPermit.tot)
specPermit$percent.revenue = specPermit$percent.revenue/specPermit$tot
specPermit.df = group_by(specPermit, taxa, specn) %>%
  filter(percent.revenue > 0.02) %>%
  arrange(taxa) %>%
  as.data.frame
specPermit.df$spec = cfec$spec[match(specPermit.df$specn, cfec$specn)]
kable(specPermit.df)
```

## Code for looking at permit
We can look at whether there's clear combinations of permit types that are combinations -- most of these people either have a single local permit (PWS, Bristol Bay, etc) or have multiple statewide permits.

```{r}

# cfecnew$area = substr(cfecnew$p_fshy,4,4)
# cfecnew$area[cfecnew$area %in% c("A","C")] = "Southeast"
# cfecnew$area[cfecnew$area == "B"] = "Statewide"
# cfecnew$area[cfecnew$area == "D"] = "Yakutat"
# cfecnew$area[cfecnew$area == "E"] = "PWS"
# cfecnew$area[cfecnew$area == "H"] = "CookInlet"
# cfecnew$area[cfecnew$area == "J"] = "Westward"
# cfecnew$area[cfecnew$area == "K"] = "Kodiak"
# cfecnew$area[cfecnew$area == "L"] = "Chignik"
# cfecnew$area[cfecnew$area == "M"] = "AlaskaPeninsula"
# cfecnew$area[cfecnew$area == "N"] = "Nelson"
# cfecnew$area[cfecnew$area == "Q"] = "BeringSea"
# cfecnew$area[cfecnew$area == "T"] = "BristolBay"
# cfecnew$area[cfecnew$area == "X"] = "Kotzebue"
# cfecnew$area[cfecnew$area == "Z"] = "NortonSound"
#
# cfecnew$spec = substr(cfecnew$p_fshy,1,1)
# cfecnew$spec[cfecnew$spec == "B"] = "Halibut"
# cfecnew$spec[cfecnew$spec == "C"] = "Sablefish"
# cfecnew$spec[cfecnew$spec %in% c("D","E","K","T")] = "Crab"
# cfecnew$spec[cfecnew$spec %in% c("F")] = "Freshwater"
# cfecnew$spec[cfecnew$spec %in% c("G","H","L")] = "Herring"
# cfecnew$spec[cfecnew$spec %in% c("I")] = "Lingcod"
# cfecnew$spec[cfecnew$spec %in% c("A","J","N","R","W")] = "Shellfish"
# cfecnew$spec[cfecnew$spec %in% c("O","P","Q","U","Z")] = "Invert"
# cfecnew$spec[cfecnew$spec %in% c("M")] = "Misc"
# cfecnew$spec[cfecnew$spec %in% c("S")] = "Salmon"
# cfecnew$area.spec = paste(cfecnew$area, cfecnew$spec, sep="_")
#
# cfecAnnual.multiplePermits = group_by(cfecnew, p_holder, year) %>%
#   summarize(npermit = length(unique(p_fshy)),
#     allpermits = paste(sort(unique(p_fshy)), collapse = " "),
#     areaspec = paste(sort(unique(area.spec)), collapse = " ")) %>%
#   filter(npermit > 1)

```
